// projects/library/tsc-lite/src/lib/tsc/enum/error.enum.ts
export enum ReportErrorEnum {
    INVALID_DATE_LABEL = 'Invalid dates!',
    INVALID_DATE_MESSAGE = 'Select distinct dates to view the desired report.',
    TIME_EXCEEDED_LABEL = 'Time exceeded!',
    HOUR_24_TIME_EXCEEDED_MESSAGE = 'You cannot view or download report for more than 24 hours.',
    DAY_30_TIME_EXCEEDED_MESSAGE = 'You cannot view or download report for more than 30 days.',
    DATA_NOT_FOUND_LABEL = 'No data found!',
    DATA_NOT_FOUND_MESSAGE = 'No data was recorded for the above specified filters. Please keep changing the filters and try again.',
    INVALID_ERROR_LABEL = 'Please try again later!',
    INVALID_ERROR_MESSAGE = 'We are not able to fetch the data for the specified filters at the moment. Please keep changing the filters and try again.'
}

export enum DashboardErrorEnum {
    DATA_NOT_FOUND_LABEL = 'No data found!',
    DATA_NOT_FOUND_MESSAGE = 'No data was recorded at the moment. Please try again after sometime.',
    INVALID_ERROR_LABEL = 'Please try again later!',
    INVALID_ERROR_MESSAGE = 'We are not able to fetch the requested data at the moment. Please try again after sometime.'
}